requirements/prod.txt
# Everything needed in production
# Flask
Flask==1.1.2
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.14.3
Flask-Moment==0.11.0
Flask-Restful==0.3.8
Flask-Marshmallow==0.14.0
Flask-JWT-Extended==4.2.1

# Additional
email-validator==1.1.2
marshmallow-sqlalchemy==0.25.0

# Security
werkzeug==1.0.1

# Enviroment
python-dotenv==0.17.0

# Deployment
gunicorn==20.1.0
